Fixes tests for Stripe 5.19.0
As of 5.19.0, the compute_signature method was made public and they
split the timestamp into a separate param (which must be a Time). This
version checks compute_signature based on the number of params it takes
since doing a version check felt odd. Now that it is public, the method
shouldn't change it's API.

1: https://github.com/stripe/stripe-ruby/pull/915